Add max kernel size for lowpass and allow temp output scaling with lowpass
Summary:
Adding max kernel height size, so we can disable lowpass filter for vertical kernel in case we need to use different anti-aliasing method. E.g. for barrel, when the output is not uniformly scaled (stretched vertically) we'd better use temporary output scale instead of lowpass vertical filtering.
Fixing barrel vertical FOV to be 90.0, since the main plain is 90 degrees vertically and not 180. Even so it's stretched, the FOV still remains 90.
Enabling intermediate output scaling even when the low pass filter is enabled to add additional smoothing and anti-aliasing when needed.

Reviewed By: puffpio

Differential Revision: D4931297

fbshipit-source-id: f6297845f12b20b27f6e1f7151aa441d344ae238