Fix seam on barrel layout.
Summary: Seam appears when outX is around 0 and when we offset it by 0.5 it gets close to left border of equirect. This results in interpolation with the transparent border and leaves the unmodified color which is black. By making outX at least half of a pixel, we remove the seam.

Differential Revision: D4946224

fbshipit-source-id: ae43468b7e7aa43cc847f854c61121bf7c346165